Specify RANLIB for android builds
When cross compiling from OSX to Android the build system for libpng would
otherwise invoke the system `ranlib` on an android-targeting archive. This
invocation would then corrupt the archive, preventing the compiler from working
with it.

By setting `RANLIB` to the android-specific `ranlib` tool the build system won't
corrupt the archive, allowing rustc to read it and continue processing it.